Reorganized queries and mutations in different folders.
