+++ src/inputs/user-input.ts
@@ -1,0 +1,6 @@
+export interface UserInput {
+  name: string;
+  email: string;
+  password: string;
+  birthdate: string;
+}
