+++ src/queries/users.ts
@@ -1,0 +1,6 @@
+import { AppDataSource } from '../data-source';
+import { User } from '../entity/User';
+
+export function getUsers() {
+  return AppDataSource.manager.getRepository(User).find();
+}
